+++ src/ExtPack_U_UART.h
@@ -35,7 +35,7 @@ static inline ext_pack_error_t send_ExtPack_UART_data(unit_t unit, uint8_t data)
  * @param send_byte_delay_us The delay between sending two bytes in us.
  * @return EXT_PACK_SUCCESS on success, EXT_PACK_FAILURE if the function was aborted when sending a uint8_t because of an error while sending.
  */
-static inline ext_pack_error_t send_ExtPack_UART_String(unit_t unit, const uint8_t* data, uint8_t send_byte_delay_us) {
+static inline ext_pack_error_t send_ExtPack_UART_String(unit_t unit, const uint8_t* data, uint16_t send_byte_delay_us) {
     return send_String_to_ExtPack(unit, data, send_byte_delay_us);
 }
